requirements.txt
aiohttp
openai
streamlit
pillow
pytesseract
python-dotenv
unstructured
pypdf
pdfkit 
